requirements.txt
selenium
imapclient
pyzmail36
pandas
psycopg2-binary
python-dotenv
